requirements.txt
pillow
idna==2.9
redis==3.4.1
requests==2.23.0
urllib3==1.25.8
pyscs
pydantic
cos-python-sdk-v5